pyproject.toml
[tool.ruff]
# Enable pycodestyle (`E`) and Pyflakes (`F`) codes by default.
select = ["E", "F", "B", "Q", "W", "C90", "N", "COM", "A", "PL"]
ignore = []

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["A", "B", "C", "D", "E", "F", "G", "I", "N", "Q", "S", "T", "W", "ANN", "ARG", "BLE", "COM", "DJ", "DTZ", "EM", "ERA", "EXE", "FBT", "ICN", "INP", "ISC", "NPY", "PD", "PGH", "PIE", "PL", "PT", "PTH", "PYI", "RET", "RSE", "RUF", "SIM", "SLF", "TCH", "TID", "TRY", "UP", "YTT"]
unfixable = []

line-length = 150

[tool.poetry]
name = "python-api"
version = "0.1.0"
description = ""
authors = ["Daniel Swarup <ucabdrs@ucl.ac.uk>"]
readme = "README.md"
packages = [{include = "src"}, {include = "tests"}]

[tool.poetry.dependencies]
python = "3.10.2"
bpy = "3.5.0"
bpycv = "0.3.6"
mathutils = "3.3.0"
numpy = "1.22.3"
opencv-python = "4.7.0.72"
pyyaml = "6.0"
requests = "2.31.0"
typer = "0.9.0"
python-dotenv = "1.0.0"
pytest = "^7.4.0"
ruff = "^0.0.282"
watchdog = "^3.0.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
